FAT-20589 rename localization file
