+++ cypress/support/fragments/settings/tenant/general/index.js
@@ -1,3 +1,3 @@
 export { default as Addresses } from './addresses';
-export { default as Localization } from './localication';
+export { default as Localization } from './localization';
 export { default as ReadingRoom } from './readingRoom';
