+++ cypress/support/fragments/settings/tenant/general/localization.js
